Crate rblas

BLAS bindings and wrappers.

Bindings are split by BLAS level and contained in a module named ll (stands for low level, not sure if that's the best name, but that's what it is).

Wrappers are split likewise. They are named after the function they wrap, but capitalized and contained in their respective ops modules. To use these wrappers, the appropriate traits must be implemented for the type. These are either Vector or Matrix.

Level 1: vector

Level 2: matrix_vector

Level 3: matrix

Reexports

pub use vector::Vector;

pub use vector::VectorOperations;

pub use matrix::Matrix;

pub use vector::ops::*;

pub use matrix_vector::ops::*;

pub use matrix::ops::*;

Modules

attribute
Various attributes of vectors and matrices.

default

math

matrix
Matrix operations.

matrix_vector
Matrix-vector operations.

vector
Vector operations.

Macros

mat!