Trait rblas::matrix::Matrix

pub trait Matrix<T> {
    fn rows(&self) -> i32;
    fn cols(&self) -> i32;
    unsafe fn as_ptr(&self) -> *const T;
    unsafe fn as_mut_ptr(&mut self) -> *mut T;

    fn lead_dim(&self) -> i32 { ... }
    fn order(&self) -> Order { ... }
}

Methods that allow a type to be used in BLAS functions as a matrix.

Required Methods

fn rows(&self) -> i32

Returns the number of rows.

fn cols(&self) -> i32

Returns the number of columns.

unsafe fn as_ptr(&self) -> *const T

An unsafe pointer to a contiguous block of memory.

unsafe fn as_mut_ptr(&mut self) -> *mut T

An unsafe pointer to a contiguous block of memory.

Provided Methods

fn lead_dim(&self) -> i32

The leading dimension of the matrix. Defaults to cols for RowMajor order and 'rows' for ColMajor order.

fn order(&self) -> Order

The order of the matrix. Defaults to RowMajor.