Trait rblas::vector::Vector

pub trait Vector<T> {
    fn len(&self) -> i32;
    unsafe fn as_ptr(&self) -> *const T;
    unsafe fn as_mut_ptr(&mut self) -> *mut T;

    fn inc(&self) -> i32 { ... }
}

Methods that allow a type to be used in BLAS functions as a vector.

Required Methods

fn len(&self) -> i32

The number of elements in the vector.

unsafe fn as_ptr(&self) -> *const T

An unsafe pointer to a contiguous block of memory.

unsafe fn as_mut_ptr(&mut self) -> *mut T

An unsafe mutable pointer to a contiguous block of memory.

Provided Methods

fn inc(&self) -> i32

The stride within the vector. For example, if inc returns 7, every 7th element is used. Defaults to 1.